API doc for log contents and misc
This commit addresses a few different things:

1. There is no trove API for log-tail and log-save.
Instead, you use the log-name utility cmd (which is not
a CLI that users use) to get the name of the swift
container that has the log. Then you use swift APIs
to get the log contents.

This commit adjusts the old log-tail description
accordingly.

2. This commit adds the volume type option to trove
create.

3. And adds a fix for signature problem on list
config defaults.

4. And adds the disable switch to log publish.

Change-Id: I3a29205be232c6cc195ed11641be6bd2c6f42485
Partial-Bug: DBAAS-974